src/main/java/com/testexample/testexample/student/StudentService.java
package com.testexample.testexample.student;

import com.testexample.testexample.student.exception.BadRequestException;
import com.testexample.testexample.student.exception.StudentNotFoundException;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.stream.Collectors;

@AllArgsConstructor
@Service
public class StudentService {

    private final StudentRepository studentRepository;

    public List<Student> getAllStudents() {
        var allStudents = studentRepository.findAll();

        return allStudents.stream().filter(x -> Gender.MALE.equals(x.getGender())).collect(Collectors.toList());
    }

    public Student addStudent(Student student) {
        Boolean existsEmail = studentRepository.selectExistsEmail(student.getEmail());
        if (existsEmail) {
            throw new BadRequestException(
                    "Email " + student.getEmail() + " taken");
        }

        return studentRepository.save(student);
    }

    public void deleteStudent(Long studentId) {
        if (!studentRepository.existsById(studentId)) {
            throw new StudentNotFoundException(
                    "Student with id " + studentId + " does not exists");
        }
        studentRepository.deleteById(studentId);
    }
}
